backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
idna==3.10
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
